refactor(web,shared,worker): improve code structure and separation of concerns
- Promote useActiveAssociation hook from features/auth to shared/hooks (used by 10+ files across 4 features)
- Extract compensation-helpers to shared/utils to eliminate cross-feature coupling
- Extract BottomNavigation and HeaderDropdown from AppShell (~280 lines, down from 504)
- Add barrel exports to assignments/index.ts for conflict detection utils
- Add ESLint rule preventing shared/ from importing feature internals (stores excluded)
- Split worker/utils.ts monolith (725 lines) into 9 focused modules with barrel re-export
- Update all import paths and test vi.mock paths accordingly
- Remove review markdown file from repository
