+++ web-app/src/features/assignments/components/AssignmentCard/AssignmentCard.tsx
@@ -2,9 +2,9 @@ import { memo, useMemo } from 'react'
 
 import type { Assignment } from '@/api/client'
 import { useAssignmentConflicts } from '@/features/assignments/hooks/useCalendarConflicts'
-import { useActiveAssociationCode } from '@/features/auth/hooks/useActiveAssociation'
 import { ExpandableCard } from '@/shared/components/ExpandableCard'
 import { features } from '@/shared/config/features'
+import { useActiveAssociationCode } from '@/shared/hooks/useActiveAssociation'
 import { useDateFormat } from '@/shared/hooks/useDateFormat'
 import { useSbbUrl } from '@/shared/hooks/useSbbUrl'
 import { useTranslation } from '@/shared/hooks/useTranslation'
