+++ web-app/src/features/exchanges/components/ExchangeFilterMenu.tsx
@@ -2,7 +2,6 @@ import { useState, useCallback, memo, useMemo } from 'react'
 
 import { useShallow } from 'zustand/react/shallow'
 
-import { useActiveAssociationCode } from '@/features/auth/hooks/useActiveAssociation'
 import {
   SlidersHorizontal,
   X,
@@ -14,6 +13,7 @@ import {
 } from '@/shared/components/icons'
 import { ResponsiveSheet } from '@/shared/components/ResponsiveSheet'
 import { features } from '@/shared/config/features'
+import { useActiveAssociationCode } from '@/shared/hooks/useActiveAssociation'
 import { useTranslation } from '@/shared/hooks/useTranslation'
 import { useSettingsStore } from '@/shared/stores/settings'
 import { formatTravelTime, MINUTES_PER_HOUR } from '@/shared/utils/format-travel-time'
