+++ web-app/src/shared/hooks/useTravelTime.ts
@@ -10,7 +10,7 @@ import { useCallback, useMemo } from 'react'
 import { useQuery, useQueryClient } from '@tanstack/react-query'
 
 import { queryKeys } from '@/api/queryKeys'
-import { useActiveAssociationCode } from '@/features/auth/hooks/useActiveAssociation'
+import { useActiveAssociationCode } from '@/shared/hooks/useActiveAssociation'
 import {
   calculateTravelTime,
   calculateMockTravelTime,
